+++ apps/block_scout_web/lib/block_scout_web/controllers/api/rpc/address_controller.ex
@@ -1,6 +1,7 @@
 defmodule BlockScoutWeb.API.RPC.AddressController do
   use BlockScoutWeb, :controller
 
+  alias BlockScoutWeb.AccessHelper
   alias BlockScoutWeb.API.RPC.Helper
   alias Explorer.{Chain, Etherscan}
   alias Explorer.Chain.{Address, Wei}
@@ -20,7 +21,7 @@ defmodule BlockScoutWeb.API.RPC.AddressController do
       |> Map.put_new(:page_number, 0)
       |> Map.put_new(:page_size, 10)
 
-    accounts = list_accounts(options)
+    accounts = list_accounts(options, AccessHelper.conn_to_ip_string(conn))
 
     conn
     |> put_status(200)
@@ -59,7 +60,7 @@ defmodule BlockScoutWeb.API.RPC.AddressController do
   def balance(conn, params, template \\ :balance) do
     with {:address_param, {:ok, address_param}} <- fetch_address(params),
          {:format, {:ok, address_hashes}} <- to_address_hashes(address_param) do
-      addresses = hashes_to_addresses(address_hashes)
+      addresses = hashes_to_addresses(address_hashes, AccessHelper.conn_to_ip_string(conn))
       render(conn, template, %{addresses: addresses})
     else
       {:address_param, :error} ->
@@ -394,20 +395,20 @@ defmodule BlockScoutWeb.API.RPC.AddressController do
     Enum.any?(address_hashes, &(&1 == :error))
   end
 
-  defp list_accounts(%{page_number: page_number, page_size: page_size}) do
+  defp list_accounts(%{page_number: page_number, page_size: page_size}, ip) do
     offset = (max(page_number, 1) - 1) * page_size
 
     # limit is just page_size
     offset
     |> Addresses.list_ordered_addresses(page_size)
-    |> trigger_balances_and_add_status()
+    |> trigger_balances_and_add_status(ip)
   end
 
-  defp hashes_to_addresses(address_hashes) do
+  defp hashes_to_addresses(address_hashes, ip) do
     address_hashes
     |> Chain.hashes_to_addresses()
-    |> trigger_balances_and_add_status()
     |> add_not_found_addresses(address_hashes)
+    |> trigger_balances_and_add_status(ip)
   end
 
   defp add_not_found_addresses(addresses, hashes) do
@@ -416,22 +417,13 @@ defmodule BlockScoutWeb.API.RPC.AddressController do
     hashes
     |> MapSet.new()
     |> MapSet.difference(found_hashes)
-    |> hashes_to_addresses(:not_found)
+    |> Enum.map(fn hash -> %Address{hash: hash, fetched_coin_balance: %Wei{value: 0}} end)
     |> Enum.concat(addresses)
   end
 
-  defp hashes_to_addresses(hashes, :not_found) do
-    Enum.map(hashes, fn hash ->
-      %Address{
-        hash: hash,
-        fetched_coin_balance: %Wei{value: 0}
-      }
-    end)
-  end
-
-  defp trigger_balances_and_add_status(addresses) do
+  defp trigger_balances_and_add_status(addresses, ip) do
     Enum.map(addresses, fn address ->
-      case CoinBalanceOnDemand.trigger_fetch(address) do
+      case CoinBalanceOnDemand.trigger_fetch(ip, address) do
         :current ->
           %{address | stale?: false}