+++ apps/explorer/lib/explorer/application.ex
@@ -285,7 +285,8 @@ defmodule Explorer.Application do
         ),
         Explorer.Migrator.RefetchContractCodes |> configure() |> configure_chain_type_dependent_process(:zksync),
         configure(Explorer.Chain.Fetcher.AddressesBlacklist),
-        Explorer.Migrator.SwitchPendingOperations
+        Explorer.Migrator.SwitchPendingOperations,
+        configure_mode_dependent_process(Explorer.Utility.RateLimiter, :api)
       ]
       |> List.flatten()