+++ apps/indexer/lib/indexer/fetcher/on_demand/token_total_supply.ex
@@ -12,14 +12,18 @@ defmodule Indexer.Fetcher.OnDemand.TokenTotalSupply do
   alias Explorer.Helper, as: ExplorerHelper
   alias Explorer.Repo
   alias Explorer.Token.MetadataRetriever
+  alias Explorer.Utility.RateLimiter
 
   @ttl_in_blocks 1
 
   ## Interface
 
-  @spec trigger_fetch(Hash.Address.t()) :: :ok
-  def trigger_fetch(address_hash) do
-    GenServer.cast(__MODULE__, {:fetch_and_update, address_hash})
+  @spec trigger_fetch(String.t() | nil, Hash.Address.t()) :: :ok
+  def trigger_fetch(caller \\ nil, address_hash) do
+    case RateLimiter.check_rate(caller, :on_demand) do
+      :allow -> GenServer.cast(__MODULE__, {:fetch_and_update, address_hash})
+      :deny -> :ok
+    end
   end
 
   ## Callbacks